resolve all bug
