+++ .env
@@ -3,4 +3,3 @@
 APP_SECRET=
 
 DATABASE_URL="sqlite:///%kernel.project_dir%/var/data.db"
-
